third_party/blink/web_tests/fast/dom/shadow/shadow-added-display-none-host-expected.txt

Tests adding a shadow root to a display none host doesn't leave dirty bits.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(inside).borderWidth is "0px"
PASS getComputedStyle(inside).borderWidth is "10px"
PASS successfullyParsed is true

TEST COMPLETE
